+++ packages/components/package-components.mjs
@@ -3,9 +3,8 @@ import path from 'path';
 import archiver from 'archiver';
 import crypto from 'crypto';
 import { spawnSync as spawn } from 'child_process';
-import { Features } from './dist/Domain/Feature/Features.js';
-
-const SOURCE_FILES_PATH = 'src/Static';
+import { Features } from '../features/dist/Domain/Feature/Features.js';
+const SOURCE_FILES_PATH = '../../node_modules';
 
 console.log('Beginning packaging procedure...');
 
@@ -16,13 +15,15 @@ if (specificFeatureIdentifier) {
 
 const TmpDir = 'tmp';
 const TmpZipDir = path.join(TmpDir);
-const ComponentsDir = path.join('components');
-const OutStaticDir = path.join(ComponentsDir);
+const ComponentsDir = path.join('all');
 
 const ChecksumsSrcPath = path.join(ComponentsDir, 'checksums.json');
 const ChecksumsDistPath = path.join(ComponentsDir, 'checksums.json');
 const Checksums = JSON.parse(fs.readFileSync(ChecksumsSrcPath).toString());
 console.log('Loaded existing checksums from', ChecksumsSrcPath);
+
+const LocationMapping = JSON.parse(fs.readFileSync('identifier-to-package.json'));
+console.log("SN ~ file: package-components.mjs ~ line 26 ~ LocationMapping", LocationMapping);
 
 function zipDirectory(sourceDir, outPath) {
   const archive = archiver('zip', { zlib: { level: 9 } });
@@ -65,14 +66,9 @@ const ensureDirExists = (dir) => {
 };
 
 const copyToTmp = async (feature) => {
-  const srcComponentPath = `${path.join(
-    SOURCE_FILES_PATH,
-    feature.identifier
-  )}`;
-  const targetComponentPath = `${path.join(
-    OutStaticDir,
-    feature.identifier
-  )}`;
+  const location = LocationMapping[feature.identifier];
+  const srcComponentPath = path.join(SOURCE_FILES_PATH, location);
+  const targetComponentPath = `${path.join(ComponentsDir, feature.identifier)}`;
 
   ensureDirExists(targetComponentPath);
 
@@ -159,7 +155,6 @@ const processFeature = async (feature) => {
 await (async () => {
   ensureDirExists(TmpDir);
   ensureDirExists(TmpZipDir);
-  ensureDirExists(OutStaticDir);
 
   const featuresToProcess = specificFeatureIdentifier
     ? [