+++ packages/snjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.31.5](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.31.4...@standardnotes/snjs@2.31.5) (2021-12-29)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+
+
+
+
+## [2.31.4](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.31.3...@standardnotes/snjs@2.31.4) (2021-12-29)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+
+
+
+
 ## [2.31.3](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.31.2...@standardnotes/snjs@2.31.3) (2021-12-29)